fix: update plan data_limit type to u32 and adjust related function signatures feat: add change_transaction_status_by_payment_intent method to TransactionRepository fix: cast plan ID to i64 in get_plan_by_id calls across handlers fix: change UserStatus enum to derive Copy trait fix: update Plan and NewPlan structs to use u32 for data_limit

+++ src/handlers/admin/plans.rs
@@ -12,7 +12,7 @@ pub async fn admin_create_plan(
     // Return the plan as JSON
     let plan = app_state
         .plan_repository()
-        .get_plan_by_id(id)
+        .get_plan_by_id(id as i64)
         .await?
         .context("Failed to get plan")?;
     Ok(Json(plan).into_response())
